benchmark/pyproject.toml
[project]
name = "benchmark"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anndata>=0.10",
    "fire>=0.5",
    "jaxtyping>=0.2",
    "joblib>=1.3",
    "numpy>=1.26",
    "pandas>=2.0",
    "polars[pyarrow]>=1.0",
    "scanpy>=1.10",
    "scipy>=1.11",
    "tqdm>=4.66",
    "hydra-core>=1.3",
]

[project.optional-dependencies]
torch = ["torch>=2.4", "torchmetrics"]
fusion = ["torch>=2.8", "lightning>=2.6"]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "ipykernel>=7.1.0",
]

[tool.setuptools.packages.find]
include = ["benchmark*"]

